feat(tests): move reload plugins into src
